refactor(backend): complete HTTP boundary extraction
